Move transport.[ch] to SP
