+++ src/core/nng_impl.h
@@ -50,7 +50,7 @@
 #include "core/url.h"
 
 // transport needs to come after url
-#include "core/transport.h"
+#include "sp/transport.h"
 
 // These have to come after the others - particularly transport.h
